use raw content, not decoded content
We never tested the Content-Encoding header.
The specification provides guidance on its use, however the order of
operations will not work right at present:

"The encoding specified by the contentEncoding keyword is independent of an
encoding specified by the Content-Type header in the request or response or
metadata of a multipart body – when both are present, the encoding specified
in the contentEncoding is applied first and then the encoding specified in the
Content-Type header."

..as we will apply the media-type decoding here in _validate_body_content,
but the contentEncoding keyword will not be applied until the JSON Schema is
applied, which comes later.

https://spec.openapis.org/oas/v3.1.0#considerations-for-file-uploads