make use of stringy_numbers feature for all parameter values
An earlier version of this feature was added way back in commit 8045b311,
which only worked when the "type" keyword was explicitly used and requested a
number, but it worked for message bodies also. Now we only support stringy
numbers when a media-type is not supplied -- if it is, it is expected that
that decoding should handle all the necessary type conversions.