track the location of all subschemas, for better $ref verification
We were previously only storing the root of each subschema, unless a more
strict metaschema dialect was being used that enforced the recursive structure
of the schema.

Also, now with JSM 0.574 we are tracking the schema locations of non-openapi
schemas as well, so these may be used interchangeably in $refs from openapi
schemas.