entirely remove JSON::Schema::Modern::Document::OpenAPI::evaluator
It is no longer treated as an attribute in JSON::Schema::Modern,
and we should solely rely on the evaluator stored by the OpenAPI::Modern
object for our uses here.

This change also fixes the data leakage (potentially between applications!)
that was just introduced in the last commit by caching documents across the
runtime, because the original document's evaluator might still be available in
the cached document object.