remove json_schema_dialect attribute from JSMDO
As per the spec (§4.8.25.5, v3.1.1):
"To allow use of a different default $schema value for all Schema Objects
contained within an OAS document, a jsonSchemaDialect value may be set within
the OpenAPI Object. If this default is not set, then the OAS dialect schema id
MUST be used for these Schema Objects."

..which means that something else (like our json_schema_dialect attribute)
cannot be used in lieu of jsonSchemaDocument in the OAD itself.