support all features with a v3.0 OpenAPI document
Includes a new JSON Schema vocabulary to implement the v3.0 JSON Schema
dialect, as peculiar as it is (not quite a superset or a subset of draft4,
with no "id" support).

All functionality should be supported now, with the exception of
embedding a 3.0 JSON Schema dialect within a 3.1/3.2 OpenAPI document,
which is technically feasible but requires a number of workarounds in various
places in the code, which is not worth implementing given its marginal
utility (just upgrade your JSON Schemas to draft2020-12!).