switch to using the more lax metaschema, allowing the use of custom "$schema" keywords
As per v3.2.0 §4.24.7, "Specifying Schema Dialects": "If [jsonSchemaDialect]
is not set, then the OAS dialect schema id MUST be used for these Schema
Objects." This implies that the use of the "$schema" keyword must remain
unrestricted by default (rather than fixed to a constant value as in the
"schema-base" metaschema.

Therefore, we no longer default to using the "schema-base" metaschema (which
restricts the use of "$schema" keywords differing from the jsonSchemaDialect),
nor do we dynamically create custom metaschemas to do the same when
jsonSchemaDialect is customized.

_dynamic_metaschema_uri() is now no longer used.